requirements.txt
blinker==1.7.0
chardet==5.2.0
click==8.1.7
Flask==3.0.2
Flask-Mail==0.9.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
reportlab==4.1.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
uuid==1.30
Werkzeug==3.0.1
